+++ include/bemos_modbus_client/version_info.hpp
@@ -1,0 +1,8 @@
+#ifndef VERSION_INFO_HPP_
+#define VERSION_INFO_HPP_
+
+constexpr auto app_version_major = 2;
+constexpr auto app_version_minor = 0;
+constexpr auto app_version_patch = 0;
+
+#endif /* VERSION_INFO_HPP_ */