+++ version_info.hpp
@@ -1,8 +1,0 @@
-#ifndef VERSION_INFO_HPP_
-#define VERSION_INFO_HPP_
-
-#define APP_VERSION_MAJOR	1
-#define APP_VERSION_MINOR	0
-#define APP_VERSION_PATCH	3
-
-#endif /* VERSION_INFO_HPP_ */