+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## 1.0.3 (unreleased)
+## 2.0.0 (12.10.2021)
+- use spdlog as logging library
+- make generic to be used in parallel with configuration files
+- client now registers data_sources if available in server and defined in configuration
+- update nlohmann/json library to v3.10.3 (https://github.com/nlohmann/json/releases/tag/v3.10.3)
+- update fmt library to v8.0.1 (https://github.com/fmtlib/fmt/releases/tag/8.0.1)
+- update spdlog library to v1.9.1 (https://github.com/gabime/spdlog/releases/tag/v1.9.1)
 
 ## 1.0.2 (20.11.2020)
 - no code changes